\name{cMonkey.data-package}
\alias{cMonkey.data-package}
\alias{cMonkey.data}
\docType{package}
\title{
Additional example data sets for use with the cMonkey integrated biclustering algorithm
}
\description{
Additional example data sets for use with the cMonkey integrated biclustering algorithm
}
\details{
\tabular{ll}{
Package: \tab cMonkey.data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-07\cr
License: \tab GPL-2\cr
}
\url{}
}
\author{
David J Reiss, Institute for Systems Biology

Maintainer: \email{dreiss@systemsbiology.org}
}
\references{
 
}
\keyword{ package }
\seealso{
\url{}


}
\examples{
 
}
