\name{mpn}
\alias{mpn}
\docType{data}
\title{
##   ## data name/kind ... ##
}
\description{
##  ## A concise (1-5 lines) description of the dataset. ##
}
\usage{data(mpn)}
\format{
  The format is:
List of 3
 $ organism: chr "mpn"
 $ k.clust : num 75
 $ ratios  : num [1:688, 1:252] 13.43 10.66 12.61 10.56 9.89 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:688] "MPN001" "MPN002" "MPN003" "MPN004" ...
  .. ..$ : chr [1:252] "X37.2degC" "Heat.Shock" "Mild.Oxidative.Stress._1" "Severe.Oxidative.Stress" ...
}
\details{
##  ## If necessary, more details than the __description__ above ##
}
\source{
##  ## reference to a publication or URL from which the data were obtained ##
}
\references{
##  ## possibly secondary sources and usages ##
}
\examples{
data(mpn)
## maybe str(mpn) ; plot(mpn) ...
}
\keyword{datasets}
